EWD-23743 - Remove extra request when delete deleted answer option
Former-commit-id: ec62519e25260386f93e3a9b292ecfe4b0a105d2

+++ sources/easygenerator.Web/App/viewmodels/questions/multipleChoice/multipleChoiceAnswers.js
@@ -39,14 +39,16 @@
                 return viewModel.clearSelection().then(function () {
                     performActionWhenAnswerIdIsSet(answer, function () {
                         viewModel.answers.remove(answer);
-                        repository.removeAnswer(questionId, ko.unwrap(answer.id)).then(function () {
-                            if (answer.isCorrect()) {
-                                setFirstAnswerCorrectness();
-                                answer.isCorrect(false);
-                            }
-                            showNotification();
+
+                        if (!answer.isDeleted())
+                            repository.removeAnswer(questionId, ko.unwrap(answer.id)).then(function () {
+                                if (answer.isCorrect()) {
+                                    setFirstAnswerCorrectness();
+                                    answer.isCorrect(false);
+                                }
+                                showNotification();
+                            });
                         });
-                    });
                 });
             };
