+++ sources/easygenerator.Web/App/viewmodels/questions/multipleSelect/multipleSelectAnswers.js
@@ -23,9 +23,11 @@
                 return viewModel.clearSelection().then(function () {
                     performActionWhenAnswerIdIsSet(answer, function () {
                         viewModel.answers.remove(answer);
-                        repository.removeAnswer(questionId, ko.unwrap(answer.id)).then(function () {
-                            showNotification();
-                        });
+
+                        if (!answer.isDeleted())
+                            repository.removeAnswer(questionId, ko.unwrap(answer.id)).then(function () {
+                                showNotification();
+                            });
                     });
                 });
             };
